requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
boto3==1.21.31
botocore==1.24.31
cached-property==1.5.2
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2.12
django-adminlte3==0.1.6
django-braces==1.15.0
django-celery==3.1.17
django-celery-beat==2.2.1
django-ckeditor==6.2.0
django-js-asset==2.0.0
django-model-utils==4.2.0
django-recaptcha==3.0.0
django-render-partial==0.4
django-star-ratings==0.9.2
django-storages==1.12.3
django-timezone-field==4.2.3
gunicorn==20.1.0
hiredis==2.0.0
idna==3.3
importlib-metadata==4.11.3
jmespath==1.0.0
kavenegar==1.1.2
kombu==5.2.4
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
pyparsing==3.0.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.2.2
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
typing_extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
